OlympicTicket/backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
django-admin-tailwind==0.0.31
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
pillow==10.3.0
PyJWT==2.8.0
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
